+++ src/app/quiz/mcq/page.tsx
@@ -1,8 +1,12 @@
 'use client';
-import { useState } from 'react';
 import Mcq from '../components/Mcq';
 import '../styles.css';
 
 export default function multiplechoice() {
-    return <Mcq></Mcq>;
+    return (
+        <Mcq
+            title="Multiple Choice Quiz"
+            description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum"
+        ></Mcq>
+    );
 }
